+++ packages/cli/src/commands/pack/intents.ts
@@ -4,7 +4,7 @@ import * as path from 'path'
 
 import BaseCommand from '../../BaseCommand'
 import { RequireScenarioFolder } from '../../utils/decorators'
-import prepareConfig from '../../utils/prepareConfig'
+import prepareConfig from '../../utils/prepare-config'
 
 const CONFIG_FILE = 'bearer.config.json'
 const HANDLER_NAME = 'index.js'
@@ -61,7 +61,7 @@ export default class PackIntents extends BaseCommand {
 
       // add handler
       this.debug(`Generated config: ${JSON.stringify(config, null, 2)}`)
-      archive.append(this.handlerContent(config), { name: HANDLER_NAME })
+      archive.append(buildIntentDeclaration(config), { name: HANDLER_NAME })
       // ZIP
 
       archive.finalize()
@@ -85,26 +85,12 @@ export default class PackIntents extends BaseCommand {
     return config
   }
 
-  handlerContent({ intents }: TConfig): string {
-    return intents
-      .map(Object.keys)
-      .map(
-        intent => `
-const ${intent} = require("./dist/${intent}").default;
-module.exports[${intent}.intentName] = ${intent}.intentType.intent(${intent}.action);
-`
-      )
-      .join('\n')
-  }
-
   // TODO: rewrite this using TS AST
-  async retrieveIntents(): Promise<Array<TIntentConfig>> {
+  async retrieveIntents(): Promise<Array<string>> {
     try {
       const config = await prepareConfig(
         this.locator.authConfigPath,
-        '',
         this.bearerConfig.scenarioUuid,
-        '',
         this.locator.srcIntentsDir
       )
       return config.intents
@@ -113,9 +99,19 @@ module.exports[${intent}.intentName] = ${intent}.intentType.intent(${intent}.act
     }
   }
 }
-type TIntentConfig = { [key: string]: string }
 
 type TConfig = {
-  intents: Array<TIntentConfig>
+  intents: Array<string>
   auth?: any
+}
+
+export function buildIntentDeclaration({ intents }: TConfig): string {
+  return intents
+    .map((intent, index) => {
+      const intentConstName = `intent${index}`
+      return `const ${intentConstName} = require("./dist/${intent}").default;
+module.exports['${intent}'] = ${intentConstName}.intentType.intent(${intentConstName}.action);
+`
+    })
+    .join('\n')
 }