+++ packages/cli/test/utils/__fixtures__/generators/type-alias.ts
@@ -1,0 +1,17 @@
+import { RetrieveState } from '@bearer/intents'
+
+type AType = { aliasedParams: string }
+
+// @ts-ignore
+type TOAUTH2AuthContext = {}
+// @ts-ignore
+type TFetchDataCallback = (p: any) => void
+
+export class TypeAliasClass {
+  // @ts-ignore
+  static intentType = RetrieveState
+
+  static action(context: TOAUTH2AuthContext, params: AType, body: any, callback: TFetchDataCallback) {
+    callback({ data: [] })
+  }
+}