+++ packages/cli/test/utils/__fixtures__/prepare-config/intents/a-wonderful-intent.ts
@@ -1,0 +1,6 @@
+import { RetrieveState } from '@bearer/intents'
+
+export default class MyIntent {
+  static intentType = RetrieveState
+  static action = () => {}
+}