+++ packages/cli/test/utils/__fixtures__/prepare-config/intents/incorrect-intent-type.ts
@@ -1,0 +1,4 @@
+export default class MyIntent {
+  static intentType = 'Something else'
+  static action = () => {}
+}