+++ packages/legacy-cli/src/lib/commands/startLocalDevelopmentServer/auth.ts
@@ -2,7 +2,7 @@ const Router = require('koa-router')
 import * as fs from 'fs'
 
 const router = new Router({ prefix: '/' })
-const authIdentifier = "123"
+const auth = '123'
 
 router.get('v1/user/initialize', ctx => {
   ctx.body = '<html><head></head><body><script>'
@@ -19,7 +19,7 @@ router.get('v1/auth/:integration_uuid', async ctx => {
       localStorage.setItem('${ctx.request.query.setupId}|${ctx.params.integration_uuid}', true)
       postRobot.send(window.opener, 'BEARER_AUTHORIZED', {
         scenarioId: '${ctx.params.integration_uuid}',
-        authIdentifier: '${authIdentifier}'
+        authId: '${auth}'
       })
       setTimeout(function() {
         window.close()