+++ packages/transpiler/__test__/__fixtures__/pre/RootComponent.tsx
@@ -3,8 +3,21 @@ import {
   BearerState,
   Intent,
   IntentType,
-  RootComponent
+  RootComponent,
+  Input,
+  Output,
+  BearerRef,
+  Event,
+  EventEmitter
 } from '@bearer/core'
+
+type Farmer = {
+  fullName: string
+}
+
+type Panda = {
+  fullPandaName: string
+}
 
 @RootComponent({
   group: 'AttachPullRequest',
@@ -13,14 +26,36 @@ import {
 export class AttachPullRequestAction {
   @Intent('ListRepositories')
   fetcher: BearerFetch
-  @RetrieveStateIntent()
-  retrieve: BearerFetch
   @BearerState()
   attachedPullRequests: Array<any>
   @BearerState({ statePropName: 'goats' })
   ducks: Array<any>
   @Intent('getPullRequest', IntentType.FetchData)
   fetchResource: BearerFetch
+  @Event() inlineDefinition: EventEmitter<{ inlineName: string; inlineNumber: number; other: Panda }>
+  @Input()
+  farmer: BearerRef<Farmer>
+  @Input({
+    group: 'SCOPE',
+    propertyReferenceIdName: 'goatId',
+    eventName: 'goatMilked',
+    intentName: 'retrieveGoat',
+    autoLoad: false
+  })
+  goat: BearerRef<{ aString: string; aNumber: number; panda: Panda }>
+
+  @Output()
+  feedPanda: BearerRef<Panda>
+
+  @Output({
+    intentName: 'manualIntentName',
+    intentReferenceIdKeyName: 'manualintentReferenceIdKeyName',
+    intentPropertyName: 'manuaLIntentPropertyName',
+    eventName: 'spongeBobdSaved',
+    propertyWatchedName: 'manualpropertyWatchedName',
+    referenceKeyName: 'manualreferenceKeyName'
+  })
+  shortPanda: BearerRef<{ inlineName: string }>
 
   render() {
     return (