+++ packages/transpiler/__test__/utils/location.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 
 export const FIXTURES = path.join(__dirname, '..', '__fixtures__')
-export const BUILD = path.join(__dirname, '..', '..', '.build/src')
+export const BUILD = path.join(__dirname, '..', '.build/src')
 
 export function UnitFixtureDirectory(dir: string) {
   return path.join(FIXTURES, 'unit', dir)