+++ packages/transpiler/__test__/utils/setup.ts
@@ -2,12 +2,13 @@ import * as fs from 'fs'
 import * as globby from 'globby'
 import * as path from 'path'
 
-const buildFolder = path.join(__dirname, '../..', '.build/src')
+const buildFolder = path.join(__dirname, '..', '.build/src')
 
 process.env.BEARER_SCENARIO_ID = 'SPONGE_BOB'
 
-globby.sync(path.join(buildFolder, '/**/*.tsx')).forEach(file => {
-  if (fs.existsSync(file)) {
-    fs.unlinkSync(file)
+globby.sync(['**/*.tsx', '**/*.ts', '**/*.json'], { cwd: buildFolder }).forEach(file => {
+  const filePath = path.join(buildFolder, file)
+  if (fs.existsSync(filePath)) {
+    fs.unlinkSync(filePath)
   }
 })