+++ packages/transpiler/src/transformers/bearer-cleaning.ts
@@ -3,7 +3,7 @@
  */
 import * as ts from 'typescript'
 
-import { Decorators } from '../constants'
+import { Decorators, Types } from '../constants'
 import { hasDecoratorNamed } from '../helpers/decorator-helpers'
 import { getNodeName } from '../helpers/node-helpers'
 import { TransformerOptions } from '../types'
@@ -19,6 +19,9 @@ export default function bearerCleaning(_options: TransformerOptions = {}): ts.Tr
           const prop = tsNode as ts.PropertyDeclaration
           if (hasDecoratorNamed(prop, Decorators.Input) || hasDecoratorNamed(prop, Decorators.Output)) {
             return null
+          }
+          if (hasDecoratorNamed(prop, Decorators.Event)) {
+            return removeEventEmitterType(prop)
           }
           break
         }
@@ -39,4 +42,19 @@ export default function bearerCleaning(_options: TransformerOptions = {}): ts.Tr
       return ts.visitEachChild(tsSourceFile, visit, _transformContext)
     }
   }
+}
+
+function removeEventEmitterType(property: ts.PropertyDeclaration): ts.PropertyDeclaration {
+  if (!property.type) {
+    return property
+  }
+  return ts.updateProperty(
+    property,
+    property.decorators,
+    property.modifiers,
+    property.name,
+    property.questionToken,
+    ts.createTypeReferenceNode(Types.EventEmitter, undefined),
+    property.initializer
+  )
 }