+++ packages/transpiler/src/types.ts
@@ -18,7 +18,7 @@ export type TComponentInputDefinition = {
   default: string | number
 }
 
-type TBasicFormat = 'string' | 'number' | 'boolean'
+type TBasicFormat = 'string' | 'number' | 'boolean' | 'object' | 'any'
 type TUnknown = 'unspecified'
 export type TOuputFormat = TUnknown | { type: TBasicFormat } | { [key: string]: TBasicFormat | Array<TBasicFormat> }