+++ packages/types/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.74.0"></a>
+# [0.74.0](https://github.com/Bearer/bearer/compare/v0.73.2...v0.74.0) (2018-12-01)
+
+
+### Features
+
+* **transpiler:** allow using existing property ([#400](https://github.com/Bearer/bearer/issues/400)) ([64f635d](https://github.com/Bearer/bearer/commit/64f635d))
+
+
+
+
+
 <a name="0.73.0"></a>
 # [0.73.0](https://github.com/Bearer/bearer/compare/v0.72.1...v0.73.0) (2018-11-16)